+++ examples/apps/branching/src/groceryList/interfaces.ts
